# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_engigu_baihu-panel_internal_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  utils.PaginationData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  vo.TaskVO:
    properties:
      agent_id:
        type: string
      clean_config:
        type: string
      command:
        type: string
      config:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      envs:
        type: string
      id:
        type: string
      languages:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      last_run:
        type: string
      name:
        type: string
      next_run:
        type: string
      random_range:
        type: integer
      retry_count:
        type: integer
      retry_interval:
        type: integer
      schedule:
        type: string
      tags:
        type: string
      timeout:
        type: integer
      trigger_type:
        type: string
      type:
        type: string
      updated_at:
        type: string
      work_dir:
        type: string
    type: object
host: localhost:8052
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Baihu Panel API Server documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Baihu Panel API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: 分页获取任务列表，支持按名称、Agent ID、标签、类型筛选
      parameters:
      - description: 任务名称
        in: query
        name: name
        type: string
      - description: Agent ID
        in: query
        name: agent_id
        type: string
      - description: 标签
        in: query
        name: tags
        type: string
      - description: 任务类型
        in: query
        name: type
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/utils.PaginationData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/vo.TaskVO'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取任务列表
      tags:
      - 任务
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: 根据 ID 删除任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除任务
      tags:
      - 任务
    get:
      consumes:
      - application/json
      description: 根据 ID 获取任务详情
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.TaskVO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取任务详情
      tags:
      - 任务
    put:
      consumes:
      - application/json
      description: 根据 ID 更新任务信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      - description: 任务更新信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/vo.TaskVO'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新任务
      tags:
      - 任务
  /tasks/stop/{logID}:
    post:
      consumes:
      - application/json
      description: 根据运行日志 ID 停止正在执行的任务
      parameters:
      - description: 运行日志ID
        in: path
        name: logID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_engigu_baihu-panel_internal_utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 停止任务
      tags:
      - 任务
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"